% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefcom2017data-package.R
\docType{data}
\name{gefcom}
\alias{gefcom}
\title{GEFCom2017 data set}
\source{
ISO New England 2017, Energy, Load, and Demand Reports, url: https://www.iso-ne.com/isoexpress/web/reports/load-and-demand/-/tree/zone-info

U.S. Office of Personnel Management 2017, Federal holidays, url: https://www.opm.gov/policy-data-oversight/pay-leave/federal-holidays/
}
\description{
Contains energy and temperature data for eight bottom level zones and two
aggregated zones for New England.
}
\keyword{data}
