% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\docType{data}
\name{root_dir}
\alias{root_dir}
\title{Process raw data}
\format{An object of class \code{character} of length 1.}
\usage{
root_dir
}
\description{
Processes the raw electricity data used in GEFCom2017. Processing involves
}
\details{
* Loading the raw excel files into data frames.
* Adding holiday and calendar variables.
* Removing daylight saving time hours.
* In leap years, shifting day of year back by one for all dates after February 28. This ensures each year has values in the DoY column.

Saves a gefcom data_frame containing raw data and calendar variables.

Author: Cameron Roach
}
\keyword{datasets}
