% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefcom2017data-package.R
\docType{data}
\name{gefcom}
\alias{gefcom}
\title{GEFCom2017 data set}
\description{
Contains energy and temperature data for eight bottom level zones and two
aggregated zones for New England.
}
\keyword{data}
